Разделение программы на модули
