+++ .gitignore
@@ -1,1 +1,3 @@
-snake
+debug/**
+*.o
+
