+++ inc/color.h
@@ -1,0 +1,8 @@
+#ifndef __COLOR_H_
+#define __COLOR_H_
+
+    #include <ncurses.h>
+
+    void setColor(int objectType);
+
+#endif
