+++ inc/help.h
@@ -1,0 +1,8 @@
+#ifndef __HELP_H_
+#define __HELP_H_
+
+    #include <ncurses.h>
+
+    void printHelp(char *s) ;
+
+#endif
