+++ src/color.c
@@ -1,0 +1,22 @@
+
+#include "../inc/color.h"
+
+void setColor(int objectType){
+    attroff(COLOR_PAIR(1));
+    attroff(COLOR_PAIR(2));
+    attroff(COLOR_PAIR(3));
+    switch (objectType){
+        case 1:{ // SNAKE1
+            attron(COLOR_PAIR(1));
+            break;
+        }
+        case 2:{ // SNAKE2
+            attron(COLOR_PAIR(2));
+            break;
+        }
+        case 3:{ // FOOD
+            attron(COLOR_PAIR(3));
+            break;
+        }
+    }
+}
