+++ src/config.c
@@ -1,0 +1,2 @@
+#include "../inc/config.h"
+
