+++ src/help.c
@@ -1,0 +1,7 @@
+#include "../inc/help.h"
+
+
+void printHelp(char *s) 
+{
+    mvprintw(0, 0, s);
+}
